Temporary build mesa against LLVM 17 in Fedora 40
LLVM 18 is in testing and it'll be available in Fedora 40 soon. While
the end goal is to build mesa against it, there are some issues reported
upstream and in Fedora's bug tracker that would need to be addressed
first.

In order to avoid issues once Fedora 40 is released, temporary build
mesa against LLVM 17 and enable LLVM 18 once the issue reports in
Fedora Rawhide stabilize a bit.