Update to v25.2.1
Add  rustc-hash dependency needed for NVK.
Since it is a direct dependency, it needs to be added
in the Cargo.toml section.

Also remove some deprecated options like gallium-xa,
gallium-nine, libselinux and libxatracker which are no
longer supported in this version of mesa

intel_clc option is now deprecated

While at it micro update the rust dependencies.

v2: Remove the deprecated options instead of disabling them.

Signed-off-by: Anusha Srivatsa <asrivats@redhat.com>